third_party/blink/web_tests/fast/css/getComputedStyle/getComputedStyle-text-decoration-expected.txt

Test to make sure text-decoration property returns values properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Initial value:
PASS computedStyle.textDecoration is 'none'


Initial value (explicit):
PASS computedStyle.textDecoration is 'none'


Value 'none':
PASS computedStyle.textDecoration is 'none'


Value 'underline':
PASS computedStyle.textDecoration is 'underline'


Value 'overline':
PASS computedStyle.textDecoration is 'overline'


Value 'line-through':
PASS computedStyle.textDecoration is 'line-through'


Value 'underline overline line-through':
PASS computedStyle.textDecoration is 'underline overline line-through'


Value 'blink' (valid but ignored):
PASS computedStyle.textDecoration is 'blink'


Value 'overline overline' (invalid):
PASS computedStyle.textDecoration is 'none'


Value 'underline blank' (invalid):
PASS computedStyle.textDecoration is 'none'


Value '':
PASS computedStyle.textDecoration is 'none'


Parent gets 'underline' value:
PASS computedStyle.textDecoration is 'underline'


Ancestor should explicitly inherit value from parent when 'inherit' value is used:
PASS computedStyle.textDecoration is 'underline'


Ancestor should not implicitly inherit value from parent (i.e. when value is void):
PASS computedStyle.textDecoration is 'none'


PASS successfullyParsed is true

TEST COMPLETE
